+++ srcs/vec.c
@@ -6,7 +6,7 @@
 /*   By: mlarboul <mlarboul@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/16 14:22:58 by mlarboul          #+#    #+#             */
-/*   Updated: 2020/12/30 19:21:45 by mlarboul         ###   ########.fr       */
+/*   Updated: 2021/01/01 20:49:47 by mlarboul         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,7 +32,7 @@ t_vec	vec_sub(t_vec v1, t_vec v2)
 	return (v);
 }
 
-t_vec	vec_mult(t_vec v1, float nb)
+t_vec	vec_mult(t_vec v1, double nb)
 {
 	t_vec	v;
 
@@ -42,7 +42,7 @@ t_vec	vec_mult(t_vec v1, float nb)
 	return (v);
 }
 
-t_vec	vec_div(t_vec v1, float nb)
+t_vec	vec_div(t_vec v1, double nb)
 {
 	t_vec	v;
 
@@ -52,7 +52,7 @@ t_vec	vec_div(t_vec v1, float nb)
 	return (v);
 }
 
-float	vec_dot(t_vec v1, t_vec v2)
+double	vec_dot(t_vec v1, t_vec v2)
 {
 	return (v1.x * v2.x + v1.y * v2.y + v1.z * v2.z);
 }
